Refactor hang dump process lifetime handler
