+++ src/Platform/Microsoft.Testing.Extensions.HangDump/HangDumpProcessLifetimeHandler.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Microsoft.Testing.Extensions.Diagnostics.Helpers;
 using Microsoft.Testing.Extensions.Diagnostics.Resources;
 using Microsoft.Testing.Extensions.HangDump.Serializers;
 using Microsoft.Testing.Platform.CommandLine;
@@ -16,11 +15,6 @@ using Microsoft.Testing.Platform.IPC.Serializers;
 using Microsoft.Testing.Platform.Logging;
 using Microsoft.Testing.Platform.Messages;
 using Microsoft.Testing.Platform.OutputDevice;
-using Microsoft.Testing.Platform.Services;
-
-#if NETCOREAPP
-using Microsoft.Diagnostics.NETCore.Client;
-#endif
 
 namespace Microsoft.Testing.Extensions.Diagnostics;
 
@@ -28,7 +22,7 @@ namespace Microsoft.Testing.Extensions.Diagnostics;
 [UnsupportedOSPlatform("ios")]
 [UnsupportedOSPlatform("tvos")]
 [UnsupportedOSPlatform("wasi")]
-internal sealed class HangDumpProcessLifetimeHandler : ITestHostProcessLifetimeHandler, IOutputDeviceDataProducer, IDataProducer,
+internal sealed partial class HangDumpProcessLifetimeHandler : ITestHostProcessLifetimeHandler, IOutputDeviceDataProducer, IDataProducer,
 #if NETCOREAPP
     IAsyncDisposable,
 #endif
@@ -193,25 +187,6 @@ internal sealed class HangDumpProcessLifetimeHandler : ITestHostProcessLifetimeH
             TimeSpan.FromMilliseconds(-1));
     }
 
-    private static string GetDiskInfo()
-    {
-        var builder = new StringBuilder();
-        DriveInfo[] allDrives = DriveInfo.GetDrives();
-
-        foreach (DriveInfo d in allDrives)
-        {
-            builder.AppendLine(CultureInfo.InvariantCulture, $"Drive {d.Name}");
-            if (d.IsReady)
-            {
-                builder.AppendLine(CultureInfo.InvariantCulture, $"  Available free space: {d.AvailableFreeSpace} bytes");
-                builder.AppendLine(CultureInfo.InvariantCulture, $"  Total free space: {d.TotalFreeSpace} bytes");
-                builder.AppendLine(CultureInfo.InvariantCulture, $"  Total size: {d.TotalSize} bytes");
-            }
-        }
-
-        return builder.ToString();
-    }
-
     public async Task OnTestHostProcessExitedAsync(ITestHostProcessInformation testHostProcessInformation, CancellationToken cancellationToken)
     {
         cancellationToken.ThrowIfCancellationRequested();
@@ -235,218 +210,6 @@ internal sealed class HangDumpProcessLifetimeHandler : ITestHostProcessLifetimeH
             await _messageBus.PublishAsync(this, new FileArtifact(new FileInfo(dumpFile), ExtensionResources.HangDumpArtifactDisplayName, ExtensionResources.HangDumpArtifactDescription)).ConfigureAwait(false);
         }
     }
-
-    [UnsupportedOSPlatform("browser")]
-    [UnsupportedOSPlatform("ios")]
-    [UnsupportedOSPlatform("tvos")]
-    [UnsupportedOSPlatform("wasi")]
-    private async Task TakeDumpOfTreeAsync(CancellationToken cancellationToken)
-    {
-        ApplicationStateGuard.Ensure(_testHostProcessInformation is not null);
-
-        await _logger.LogInformationAsync($"Hang dump timeout({_activityTimerValue}) expired.").ConfigureAwait(false);
-        await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.HangDumpTimeoutExpired, _activityTimerValue)), cancellationToken).ConfigureAwait(false);
-
-        using IProcess process = _processHandler.GetProcessById(_testHostProcessInformation.PID);
-        var processTree = (await process.GetProcessTreeAsync(_logger, _outputDisplay, cancellationToken).ConfigureAwait(false)).Where(p => p.Process?.Name is not null and not "conhost" and not "WerFault").ToList();
-        IEnumerable<IProcess> bottomUpTree = processTree.OrderByDescending(t => t.Level).Select(t => t.Process).OfType<IProcess>();
-
-        try
-        {
-            if (processTree.Count > 1)
-            {
-                await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(ExtensionResources.DumpingProcessTree), cancellationToken).ConfigureAwait(false);
-
-                foreach (ProcessTreeNode? p in processTree.OrderBy(t => t.Level))
-                {
-                    await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData($"{(p.Level != 0 ? " + " : " > ")}{new string('-', p.Level)} {p.Process!.Id} - {p.Process.Name}"), cancellationToken).ConfigureAwait(false);
-                }
-            }
-            else
-            {
-                await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.DumpingProcess, process.Id, process.Name)), cancellationToken).ConfigureAwait(false);
-            }
-
-            await _logger.LogInformationAsync($"Hang dump timeout({_activityTimerValue}) expired.").ConfigureAwait(false);
-
-            // Do not suspend processes with NetClient dumper it stops the diagnostic thread running in
-            // them and hang dump request will get stuck forever, because the process is not co-operating.
-            // Instead we start one task per dump asynchronously, and hope that the parent process will start dumping
-            // before the child process is done dumping. This way if the parent is waiting for the children to exit,
-            // we will be dumping it before it observes the child exiting and we get a more accurate results. If we did not
-            // do this, then parent that is awaiting child might exit before we get to dumping it.
-            foreach (IProcess p in bottomUpTree)
-            {
-                try
-                {
-                    await TakeDumpAsync(p, cancellationToken).ConfigureAwait(false);
-                }
-                catch (Exception e)
-                {
-                    // exceptions.Add(new InvalidOperationException($"Error while taking dump of process {p.Name} {p.Id}", e));
-                    await _logger.LogErrorAsync($"Error while taking dump of process {p.Id} - {p.Name}", e).ConfigureAwait(false);
-                    await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.ErrorWhileDumpingProcess, p.Id, p.Name, e)), cancellationToken).ConfigureAwait(false);
-                }
-            }
-        }
-        finally
-        {
-            NotifyCrashDumpServiceIfEnabled();
-
-            // Some of the processes might crashed, which breaks the process tree (on windows it is just an illusion),
-            // so try extra hard to kill all the known processes in the tree, since we already spent a bunch of time getting
-            // to know which processes are involved.
-            foreach (ProcessTreeNode node in processTree)
-            {
-                IProcess? p = node.Process;
-                if (p == null)
-                {
-                    continue;
-                }
-
-                try
-                {
-                    if (!p.HasExited)
-                    {
-                        p.Kill();
-                        await p.WaitForExitAsync(CancellationToken.None).ConfigureAwait(false);
-                    }
-                }
-                catch (Exception e)
-                {
-                    await _logger.LogErrorAsync($"Problem killing {p.Id} - {p.Name}", e).ConfigureAwait(false);
-                    await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.ErrorKillingProcess, p.Id, p.Name, e)), cancellationToken).ConfigureAwait(false);
-                }
-            }
-        }
-    }
-
-    private async Task TakeDumpAsync(IProcess process, CancellationToken cancellationToken)
-    {
-        ApplicationStateGuard.Ensure(_testHostProcessInformation is not null);
-        ApplicationStateGuard.Ensure(_dumpType is not null);
-
-        string processId = process.Id.ToString(CultureInfo.InvariantCulture);
-        Dictionary<string, string> replacements = ArtifactNamingHelper.GetStandardReplacements(process.Name, processId, _clock.UtcNow);
-
-        string pattern = _dumpFileNamePattern ?? $"{process.Name}_%p_hang.dmp";
-
-        // First resolve {placeholder} templates, then handle legacy %p pattern for backward compatibility.
-        string finalDumpFileName = ArtifactNamingHelper.ResolveTemplate(pattern, replacements)
-            .Replace("%p", processId);
-        string resultsDirectory = Path.GetFullPath(_configuration.GetTestResultDirectory());
-        finalDumpFileName = Path.GetFullPath(Path.Combine(resultsDirectory, finalDumpFileName));
-
-        // Reject resolved paths that escape the results directory (e.g. rooted paths or ".." segments).
-        // Append a trailing separator to prevent sibling-directory bypass (e.g. "/tmp/results" vs "/tmp/results-evil").
-        // Use case-insensitive comparison on Windows where paths are case-insensitive.
-        StringComparison pathComparison = RuntimeInformation.IsOSPlatform(OSPlatform.Windows)
-            ? StringComparison.OrdinalIgnoreCase
-            : StringComparison.Ordinal;
-        string separatorStr = Path.DirectorySeparatorChar.ToString();
-        string resultsDirectoryGuard = resultsDirectory.EndsWith(separatorStr, StringComparison.Ordinal)
-            ? resultsDirectory
-            : resultsDirectory + separatorStr;
-        if (!finalDumpFileName.StartsWith(resultsDirectoryGuard, pathComparison))
-        {
-            throw new InvalidOperationException($"The resolved dump file path '{finalDumpFileName}' is outside the results directory '{resultsDirectory}'. Ensure --hangdump-filename is a relative path without '..' segments.");
-        }
-
-        // Ensure the destination directory exists (templates may include directory separators, e.g. {asm}/{pname}).
-        Directory.CreateDirectory(Path.GetDirectoryName(finalDumpFileName)!);
-
-        ApplicationStateGuard.Ensure(_namedPipeClient is not null);
-        GetInProgressTestsResponse tests = await _namedPipeClient.RequestReplyAsync<GetInProgressTestsRequest, GetInProgressTestsResponse>(new GetInProgressTestsRequest(), cancellationToken).ConfigureAwait(false);
-        if (tests.Tests.Length > 0)
-        {
-            string hangTestsFileName = Path.ChangeExtension(finalDumpFileName, ".log");
-            using (FileStream fs = File.OpenWrite(hangTestsFileName))
-            using (StreamWriter sw = new(fs))
-            {
-                await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(ExtensionResources.RunningTestsWhileDumping), cancellationToken).ConfigureAwait(false);
-                foreach ((string testName, int seconds) in tests.Tests)
-                {
-                    await sw.WriteLineAsync($"[{TimeSpan.FromSeconds(seconds)}] {testName}").ConfigureAwait(false);
-                    await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData($"[{TimeSpan.FromSeconds(seconds)}] {testName}"), cancellationToken).ConfigureAwait(false);
-                }
-            }
-
-            await _messageBus.PublishAsync(this, new FileArtifact(new FileInfo(hangTestsFileName), ExtensionResources.HangTestListArtifactDisplayName, ExtensionResources.HangTestListArtifactDescription)).ConfigureAwait(false);
-        }
-
-        await _logger.LogInformationAsync($"Creating dump filename {finalDumpFileName}").ConfigureAwait(false);
-
-        await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.CreatingDumpFile, finalDumpFileName)), cancellationToken).ConfigureAwait(false);
-
-#if NETCOREAPP
-        DiagnosticsClient diagnosticsClient = new(process.Id);
-        DumpType? dumpType = _dumpType.ToLowerInvariant().Trim() switch
-        {
-            "mini" => DumpType.Normal,
-            "heap" => DumpType.WithHeap,
-            "triage" => DumpType.Triage,
-            "full" => DumpType.Full,
-            "none" => null,
-            _ => throw ApplicationStateGuard.Unreachable(),
-        };
-
-        DumpFileNames dumpFileNames = GetDumpFileNames(finalDumpFileName);
-
-        try
-        {
-            // Skip creating the dump if the option is set to none, and just kill the process.
-            if (dumpType.HasValue)
-            {
-                diagnosticsClient.WriteDump(dumpType.Value, dumpFileNames.WriteDumpFileName, logDumpGeneration: false);
-                _dumpFiles.Add(dumpFileNames.ArtifactDumpFileName);
-            }
-        }
-        catch (Exception e)
-        {
-            await _logger.LogErrorAsync($"Error while writing dump of process {process.Name} {process.Id}", e).ConfigureAwait(false);
-            await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.ErrorWhileDumpingProcess, process.Id, process.Name, e)), cancellationToken).ConfigureAwait(false);
-        }
-
-#else
-        MiniDumpWriteDump.MiniDumpTypeOption? miniDumpTypeOption = _dumpType.ToLowerInvariant().Trim() switch
-        {
-            "mini" => MiniDumpWriteDump.MiniDumpTypeOption.Mini,
-            "heap" => MiniDumpWriteDump.MiniDumpTypeOption.Heap,
-            "full" => MiniDumpWriteDump.MiniDumpTypeOption.Full,
-            "none" => null,
-            _ => throw ApplicationStateGuard.Unreachable(),
-        };
-
-        try
-        {
-            // Skip creating the dump if the option is set to none, and just kill the process.
-            if (miniDumpTypeOption.HasValue)
-            {
-                MiniDumpWriteDump.CollectDumpUsingMiniDumpWriteDump(process.Id, finalDumpFileName, miniDumpTypeOption.Value);
-                _dumpFiles.Add(finalDumpFileName);
-            }
-        }
-        catch (Exception e)
-        {
-            await _logger.LogErrorAsync($"Error while writing dump of process {process.Name} {process.Id}", e).ConfigureAwait(false);
-            await _outputDisplay.DisplayAsync(new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.ErrorWhileDumpingProcess, process.Id, process.Name, e)), cancellationToken).ConfigureAwait(false);
-        }
-#endif
-    }
-
-    // Wrap the dump path into "" when it has space in it, this is a workaround for this runtime issue: https://github.com/dotnet/diagnostics/issues/5020
-    // It only affects windows. Otherwise the dump creation fails with: [createdump] The pid argument is no longer supported
-    internal static DumpFileNames GetDumpFileNames(string dumpFileName)
-        => new(
-            RuntimeInformation.IsOSPlatform(OSPlatform.Windows) && dumpFileName.Contains(' ')
-                ? $"\"{dumpFileName}\""
-                : dumpFileName,
-            dumpFileName);
-
-    internal readonly record struct DumpFileNames(string WriteDumpFileName, string ArtifactDumpFileName);
-
-    private static void NotifyCrashDumpServiceIfEnabled()
-        => AppDomain.CurrentDomain.SetData("ProcessKilledByHangDump", "true");
 
     public void Dispose()
     {
@@ -495,25 +258,4 @@ internal sealed class HangDumpProcessLifetimeHandler : ITestHostProcessLifetimeH
         _singleConnectionNamedPipeServer?.Dispose();
     }
 #endif
-}
-
-internal sealed class OutputDeviceWriter
-{
-    private readonly IOutputDevice _outputDevice;
-    private readonly IOutputDeviceDataProducer _outputDeviceDataProducer;
-
-    public OutputDeviceWriter(IOutputDevice outputDevice, IOutputDeviceDataProducer outputDeviceDataProducer)
-    {
-        _outputDevice = outputDevice;
-        _outputDeviceDataProducer = outputDeviceDataProducer;
-    }
-
-    /// <summary>
-    /// Displays the output data asynchronously, using the stored producer.
-    /// </summary>
-    /// <param name="data">The output data.</param>
-    /// <param name="cancellationToken">The cancellation token.</param>
-    /// <returns>A task representing the asynchronous operation.</returns>
-    public async Task DisplayAsync(IOutputDeviceData data, CancellationToken cancellationToken)
-        => await _outputDevice.DisplayAsync(_outputDeviceDataProducer, data, cancellationToken).ConfigureAwait(false);
 }
